mnist_project/requirements.txt
fastapi
uvicorn
streamlit
streamlit-drawable-canvas
Pillow
numpy
torch
torchvision
requests
python-multipart
opencv-python-headless
